Collect and summarize results from MB-sPLS nested CV Batchtools jobs

This function collects and summarizes the results from the MB-sPLS nested cross-validation jobs that were submitted to the Batchtools registry. It gathers the results from each outer fold, including the final model performance metrics and the inner cross-validation results. It also computes summary statistics across all outer folds.

collect_mbspls_nested_cv(ids = NULL, reg)

Arguments

ids

A vector of job IDs corresponding to the completed outer fold jobs.

reg

A batchtools registry object returned by mbspls_nested_cv_batchtools().

Value

A list containing:

results: A data.table with the results from each outer fold.

c_mats: A list of the tuned C* matrices from each outer fold.

inner_scores: A numeric vector of the best inner cross-validation scores from each outer fold.

payloads: (optional) A list of the full payloads from each outer fold (if store_payload = TRUE was set).

summary_table: A data.table with summary statistics across all outer folds.

See also